move methods from mute service to chat service
