+++ src/test/java/com/faforever/client/chat/AbstractChatTabControllerTest.java
@@ -104,8 +104,6 @@ public class AbstractChatTabControllerTest extends UITest {
   private ChatUserService chatUserService;
   @Mock
   private EmoticonService emoticonService;
-  @Mock
-  private MuteService muteService;
 
   private AbstractChatTabController instance;
   private CountDownLatch chatReadyLatch;
@@ -127,7 +125,7 @@ public class AbstractChatTabControllerTest extends UITest {
 
     instance = new AbstractChatTabController(webViewConfigurer, userService, chatService, preferencesService,
         playerService, audioService, timeService, i18n, imageUploadService, notificationService, reportingService,
-        uiService, eventBus, countryFlagService, chatUserService, emoticonService, muteService) {
+        uiService, eventBus, countryFlagService, chatUserService, emoticonService) {
       private final Tab root = new Tab();
       private final WebView webView = new WebView();
       private final TextInputControl messageTextField = new TextField();
